Use official IDisposable pattern in CanvasEffect
